* Reapply some patchs on SQLObject that we need for Stoq to work properly. A
  lot of them were not necessary, as they were fixed upstream.
* Correct Stoq code at some places to fix some problems that appeared with the
  upgrade, specially Viewable and it's subclasses.

r=jdahlin,romaia

+++ external/sqlobject/col.py
@@ -763,7 +763,7 @@ class FloatCol(Col):
 
 
 class SOKeyCol(SOCol):
-    key_type = {int: "INT", str: "TEXT"}
+    key_type = {int: "BIGINT", str: "TEXT"}
 
     # 3-03 @@: this should have a simplified constructor
     # Should provide foreign key information for other DBs.
